requirements.txt
alembic==1.7.6
anyio==3.6.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
cliapp==1.0.9
click==8.0.3
cloudinary==1.29.0
colorama==0.4.4
command-not-found==0.3
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.51.0
django-bootstrap3==21.2
django-bootstrap4==22.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-heroku==0.3.1
django-otp==1.1.3
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-registration==3.2
django-registration-redux==2.10
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-tinymce==3.4.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
drf-yasg==1.20.0
duplicity==0.8.21
email-validator==1.2.1
fasteners==0.14.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
gyp==0.1
h11==0.12.0
httpcore==0.13.7
httplib2==0.20.2
httpx==0.18.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
phonenumbers==8.12.51
Pillow==9.0.1
protobuf==3.12.4
psycopg2==2.9.3
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyaml==21.10.1
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pydantic==1.9.1
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
pyuploadcare==3.0.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.6
whitenoise==6.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
